Check for empty strings rather than falsey values
In Ruby, an empty string is truthy. As a result, passed a value of "" or
"nil", these branches of the code will run.

As a result, we need to change the default values of the input to `""`
and check if the passed value is an empty string.

No parsing is done to the values passed in as args so these are just
strings. `"nil"` doesn't become `nil` and is truthy.